+++ src/mol-gl/shader/cylinders.vert.ts
@@ -57,9 +57,6 @@ void main() {
     // ensure cylinder 'dir' is pointing towards the camera
     if(dot(camDir, dir) < 0.0) {
         dir = -dir;
-        vec3 tmp = vStart;
-        vStart = vEnd;
-        vEnd = tmp;
     }
 
     vec3 left = cross(camDir, dir);